Add db contex, flowers controller children controller and attribute validation

+++ KindergartenApp/Models/Child.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -14,12 +15,14 @@ namespace KindergartenApp.Models
     public class Child
     {
         public long Id { get; set; }
+       // [Required]
+        //[MaxLength(10, ErrorMessage = "Name must have almost 10 characters ")]
         public string Name { get; set; }
         public string Gender { get; set; }
         public string Characterization { get; set; }
         public double Age { get; set; }
         public long Height { get; set; }
-        public long Weight { get; set; }
+        public double Weight { get; set; }
         public ChildrenGroup ChildrenGroup { get; set; }
     }
 }
